JAMES-1902 Renames activemq mailqueue guice bindings module
