+++ server/container/guice/queue/activemq/pom.xml
@@ -26,7 +26,7 @@
         <relativePath>../../pom.xml</relativePath>
     </parent>
 
-    <artifactId>james-server-guice-activemq</artifactId>
+    <artifactId>queue-activemq-guice</artifactId>
 
     <name>Apache James :: Server :: Guice :: ActiveMQ</name>
     <description>Guice Module for ActiveMQ</description>
